Give <html> tags `lang="en"` attributes
In `/index.html` and `/util/index.html`, give the root `<html>` tags the
`lang="en"` attribute.

A run of <http://validator.w3.org/#validate_by_input> by @nyuszika7h on
<https://8573.github.io> complained that `lang` is required if
`xml:lang` is present on an element “in no namespace”. Of course this
element *is* in a namespace, but I’m adding this anyway in the spirit of
defensiveness.

I note that <http://validator.w3.org/#validate_by_uri> does not complain
about this or any other issue when run on <https://8573.github.io>.